core/requirements.txt
python-binance
pandas
ccxt
fastapi
uvicorn
matplotlib
fastapi
flask
dash
plotly
tradingpattern
colorama
python-telegram-bot==13.13
websocket-client
ta